Rename files so module names are all lowercase
